confirmed python validation tests
